+++ application/templates/feed.html
@@ -265,7 +265,7 @@
                             <h5 id="name-contact-{{ item.item_id }}"></h5>
                             <h5 id="email-contact-{{ item.item_id }}"></h5>
                             <form name="Message" method="post" class="form mt-4"
-                                  action="{{ url_for('send_message', recipient=item.user_id) }}">
+                                  action="{{ url_for('message.send_message', recipient=item.user_id) }}">
                                 {{ send_message_form.hidden_tag() }}
                                 <div class="form-group">
                                     <label for="message"><h5><strong>Send Message</strong></h5></label>
