+++ application/templates/messages.html
@@ -80,7 +80,7 @@
                                 <h5><strong>Item:&nbsp;&nbsp;&nbsp;</strong>Lost snek</h5>
 
                                 <form name="Message" method="post" class="form mt-4"
-                                      action="{{ url_for('send_message', recipient=msg.author.user_id) }}">
+                                      action="{{ url_for('message.send_message', recipient=msg.author.user_id) }}">
                                     {{ send_message_form.hidden_tag() }}
                                     <div class="form-group">
                                         <label for="message"><h5><strong>Message</strong></h5></label>
